+++ news matters/Helper/UIImageView+Extensions.swift
@@ -1,0 +1,30 @@
+//
+//  UIImageView+Extensions.swift
+//  news matters
+//
+//  Created by Muhammad Ali Maniar on 20/11/2021.
+//
+
+import Foundation
+import UIKit
+
+extension UIImageView {
+    func downloaded(from url: URL, contentMode mode: ContentMode = .scaleAspectFit) {
+        contentMode = mode
+        URLSession.shared.dataTask(with: url) { data, response, error in
+            guard
+                let httpURLResponse = response as? HTTPURLResponse, httpURLResponse.statusCode == 200,
+                let mimeType = response?.mimeType, mimeType.hasPrefix("image"),
+                let data = data, error == nil,
+                let image = UIImage(data: data)
+                else { return }
+            DispatchQueue.main.async() { [weak self] in
+                self?.image = image
+            }
+        }.resume()
+    }
+    func downloaded(from link: String?, contentMode mode: ContentMode = .scaleAspectFill) {
+        guard let link = link, let url = URL(string: link) else { return }
+        downloaded(from: url, contentMode: mode)
+    }
+}
