+++ news matters/Presentation/NewsView/NewsCell/NewsCell.swift
@@ -29,34 +29,3 @@ class NewsCell: UITableViewCell {
     }
     
 }
-
-extension UIView {
-    func createCardView() {
-        layer.cornerRadius = 20.0
-        layer.shadowColor = UIColor.darkGray.cgColor
-        layer.shadowOffset = CGSize(width: 0.0, height: 0.0)
-        layer.shadowRadius = 8.0
-        layer.shadowOpacity = 0.7
-    }
-}
-
-extension UIImageView {
-    func downloaded(from url: URL, contentMode mode: ContentMode = .scaleAspectFit) {
-        contentMode = mode
-        URLSession.shared.dataTask(with: url) { data, response, error in
-            guard
-                let httpURLResponse = response as? HTTPURLResponse, httpURLResponse.statusCode == 200,
-                let mimeType = response?.mimeType, mimeType.hasPrefix("image"),
-                let data = data, error == nil,
-                let image = UIImage(data: data)
-                else { return }
-            DispatchQueue.main.async() { [weak self] in
-                self?.image = image
-            }
-        }.resume()
-    }
-    func downloaded(from link: String?, contentMode mode: ContentMode = .scaleAspectFill) {
-        guard let link = link, let url = URL(string: link) else { return }
-        downloaded(from: url, contentMode: mode)
-    }
-}
